backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
kafka-python>=2.0.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
